+++ BranchSDK/BNCSKAdNetwork.h
@@ -10,6 +10,13 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+typedef NS_ENUM(NSInteger, BranchSkanWindow) {
+    BranchSkanWindowInvalid = 0,
+    BranchSkanWindowFirst = 1,
+    BranchSkanWindowSecond = 2,
+    BranchSkanWindowThird = 3
+};
+
 @interface BNCSKAdNetwork : NSObject
 
 @property (nonatomic, assign, readwrite) NSTimeInterval maxTimeSinceInstall;
@@ -22,6 +29,21 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)updatePostbackConversionValue:(NSInteger)conversionValue
                     completionHandler:(void (^)(NSError *error))completion;
+
+- (void)updatePostbackConversionValue:(NSInteger)fineValue
+                          coarseValue:(NSString *) coarseValue
+                           lockWindow:(BOOL)lockWindow
+                    completionHandler:(void (^)(NSError *error))completion API_AVAILABLE(ios(16.1));
+
+- (int) calculateSKANWindowForTime:(NSDate *) currentTime;
+
+- (NSString *) getCoarseConversionValueFromDataResponse:(NSDictionary *) dataResponseDictionary;
+
+- (BOOL) getLockedStatusFromDataResponse:(NSDictionary *) dataResponseDictionary;
+
+- (BOOL) getAscendingOnlyFromDataResponse:(NSDictionary *) dataResponseDictionary;
+
+- (BOOL) shouldCallPostbackForDataResponse:(NSDictionary *) dataResponseDictionary;
 
 @end